% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{df_to_h5}
\alias{df_to_h5}
\title{Data frame to h5}
\usage{
df_to_h5(df, h5, gr_name = NULL)
}
\arguments{
\item{df}{Data frame of cell annotation or gene annotation}

\item{h5}{The h5 file}

\item{gr_name}{The group name represents the property of the data frame.}
}
\description{
Data frame is converted to and saved the h5 file
}
