% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{h5_to_df}
\alias{h5_to_df}
\title{H5 to dataframe}
\usage{
h5_to_df(h5df)
}
\arguments{
\item{h5df}{The hdf5 group saved the data frame using list mode.}
}
\value{
A data frame(annotation)
}
\description{
H5 gruop is converted to usable data frame including the observes annotation (cell annotation) and variables annotation (gene annotation)
}
