% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{h5_to_matrix}
\alias{h5_to_matrix}
\title{H5 to Matrix format}
\usage{
h5_to_matrix(h5mat, obs.name = NULL, var.name = NULL)
}
\arguments{
\item{h5}{The name of the group the stores the matrix data in h5 file}

\item{obs_names}{The observes names, such as cell names}

\item{var_names}{The variables names, such as gene names}
}
\value{
dense matrix or sparse matrix
}
\description{
H5 group is converted to usable matrice including dense matrices and sparse matrices.
Dense matrices(by R) and sparse matrices(constructed by Matrix package)
}
