% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{h5_to_seurat}
\alias{h5_to_seurat}
\title{H5 to the seurat}
\usage{
h5_to_seurat(h5, assay.name)
}
\arguments{
\item{h5}{The h5 file in R}

\item{assay.name}{'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
H5 file is converted to the seurat obejct
}
