% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{h5_to_spatial}
\alias{h5_to_spatial}
\title{The h5 group spatial to the spatial message}
\usage{
h5_to_spatial(h5spa)
}
\arguments{
\item{h5spa}{The h5 group for the spatial}
}
\value{
The spatial message
}
\description{
The h5 group spatial to the spatial message
}
