% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{matrix_to_h5}
\alias{matrix_to_h5}
\title{Matrix to H5 format}
\usage{
matrix_to_h5(
  mat,
  h5,
  gr_name = NULL,
  save.obs.name = FALSE,
  save.var.name = FALSE
)
}
\arguments{
\item{mat}{The matrix object including matrix(R) and sparse matrix(Matrix package)}

\item{h5}{The H5 file name that we write in}

\item{gr_name}{The h5 gorup store the matrix (dense matrix or sparse matrix)}

\item{save.obs.name}{The rownames isn't be saved(FALSE by defualt)}

\item{save.var.name}{The colnames isn't be saved(FALSE by defualt)}
}
\description{
The matrix including the dense matrix and sparse matrix is converted to the matrix in h5 format or is stored into the h5 file.
}
