% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{read_h5}
\alias{read_h5}
\title{H5 to scRNAs-seq analysis object}
\usage{
read_h5(file, target.object = "seurat", assay.name = "RNA")
}
\arguments{
\item{file}{The h5 file}

\item{target.object}{Denotes which object to load. Defualt is "seurat". Available options are:
\itemize{
  \item "seurat": converted the h5 file to "seurat object".
  \item "singlecellexperiment": converted the h5 file to "singlecellexperiment object".
  \item "monocle": converted the h5 file to "monocle3 object".
}}

\item{assay.name}{'assay.name' is used to flag the data type. Defualt is "RNA". Available options are:
\itemize{
  \item "RNA": this is the scRNA-seq data.
  \item "spatial": this is the spatial data.}}
}
\value{
The single cell analysis mainstream software
}
\description{
Read h5 and converted h5 to the scRNA-seq analysis object
}
