% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{read_h5ad}
\alias{read_h5ad}
\title{reading the h5ad file}
\usage{
read_h5ad(file, assay_name, target.object)
}
\arguments{
\item{file}{the file name}

\item{assay_name}{the assay_name}

\item{target.object}{the target.object name}
}
\description{
reading the h5ad file
}
