% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{read_h5part}
\alias{read_h5part}
\title{The read the some data in the h5 fiel}
\usage{
read_h5part(h5, groups)
}
\arguments{
\item{h5}{the h5 file}

\item{groups}{the h5 groups}
}
\description{
The read the some data in the h5 fiel
}
