% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceIO.R
\name{sce_to_h5}
\alias{sce_to_h5}
\title{The singlecellexperiment is converted to h5 file}
\usage{
sce_to_h5(sce, h5, assay.name)
}
\arguments{
\item{sce}{The singlecellexperiment object.}

\item{h5}{The h5 file in R.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
Singlecellexperiment object is converted to h5 file.
}
