% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceIO.R
\name{sce_write_h5}
\alias{sce_write_h5}
\title{The singlecellexperiment is converted to h5 file}
\usage{
sce_write_h5(sce = NULL, file = NULL, assay.name = NULL)
}
\arguments{
\item{sce}{The singlecellexperiment object.}

\item{file}{The h5 file.}

\item{assay.name}{The 'assay.name' is used to flag the data type, and the default is "RNA", meaning this is scRNA-seq data.}
}
\description{
Singlecellexperiment object is converted to h5 file.
}
