% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat.insert_dimr_}
\alias{seurat.insert_dimr_}
\title{seurat insert the dimr information}
\usage{
seurat.insert_dimr_(seurat, seurat_list)
}
\arguments{
\item{seurat}{The seruat object}

\item{seurat_list}{The list will be inserted to the seurat}
}
\description{
seurat insert the dimr information
}
