% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat.insert_graphs_}
\alias{seurat.insert_graphs_}
\title{seurat insert the graphs information}
\usage{
seurat.insert_graphs_(seurat, seurat_list)
}
\arguments{
\item{seurat}{The seruat object}

\item{seurat_list}{The list will be inserted to the seurat}
}
\description{
seurat insert the graphs information
}
