% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat.insert_uns_}
\alias{seurat.insert_uns_}
\title{seurat insert the uns information}
\usage{
seurat.insert_uns_(seurat, seurat_list)
}
\arguments{
\item{seurat}{The seruat object}

\item{seurat_list}{The list will be inserted to the seurat}
}
\description{
seurat insert the uns information
}
