% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{seurat.to_dimr_}
\alias{seurat.to_dimr_}
\title{h5[['dimR']] file transform the dimr information of seurat object}
\usage{
seurat.to_dimr_(h5)
}
\arguments{
\item{h5}{The h5 file}
}
\description{
h5[['dimR']] file transform the dimr information of seurat object
}
