% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{seurat.to_spatial_}
\alias{seurat.to_spatial_}
\title{h5[['spatial']] file transforma the spatial information of seurat object}
\usage{
seurat.to_spatial_(h5)
}
\arguments{
\item{h5}{The h5 file}
}
\description{
h5[['spatial']] file transforma the spatial information of seurat object
}
