% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat_spatial_to_h5}
\alias{seurat_spatial_to_h5}
\title{The spatial meassage to the h5 file(Seurat)}
\usage{
seurat_spatial_to_h5(data, h5, gr_name = "spatial")
}
\arguments{
\item{data}{The seruat object}

\item{h5}{The h5 file}

\item{assay.name}{'assay.name' must be the 'spatial' for saving the spatial message.}
}
\description{
The spatial meassage to the h5 file(Seurat)
}
