% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat_to_h5}
\alias{seurat_to_h5}
\title{Seurat is converted to h5 file}
\usage{
seurat_to_h5(
  seurat = NULL,
  h5 = NULL,
  assay.name = NULL,
  save.graphs = FALSE,
  save.scale = FALSE
)
}
\arguments{
\item{seurat}{The seurat object.}

\item{h5}{The h5 file in R.}

\item{assay.name}{'assay.name' is used to flag the data type. Defualt is "RNA". Available options are:
\itemize{
  \item "RNA": this is the scRNA-seq data.
  \item "spatial": this is the spatial data.}}

\item{save.graphs}{Default is False , determing whether to save the graph(cell-cell similarity network).
seurat graph is different from scanpy graph. Their relationship are set {"distances": "knn", "connectivities": "snn"} roughly.}
}
\description{
Seurat object is converted to h5 file.
}
