% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\name{seurat_write_h5}
\alias{seurat_write_h5}
\title{The seurat object to h5}
\usage{
seurat_write_h5(
  seurat = NULL,
  file = NULL,
  assay.name = NULL,
  save.graphs = TRUE,
  save.scale = FALSE
)
}
\arguments{
\item{seurat}{The seurat object.}

\item{file}{The h5 file.}

\item{assay.name}{'assay.name' is used to flag the data type. Defualt is "RNA". Available options are:
\itemize{
  \item "RNA": this is the scRNA-seq data.
  \item "spatial": this is the spatial data.}}

\item{save.graphs}{Default is TRUE , determing whether to save the graph(cell-cell similarity network).
seurat graph is different from scanpy graph. Their relationship are set {"distances": "knn", "connectivities": "snn"} roughly.}

\item{save.scale}{Default is FALSE, determint whether to save the scale.data(dense matrix)}
}
\description{
The Seurat object is converted to the h5 file.
}
