% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{to_dimr_}
\alias{to_dimr_}
\title{h5[['dimR']] file transform the dimr information}
\usage{
to_dimr_(h5)
}
\arguments{
\item{h5}{The h5 file}
}
\description{
h5[['dimR']] file transform the dimr information
}
