% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{to_graphs_}
\alias{to_graphs_}
\title{h5[['graphs']] file transform graphs information}
\usage{
to_graphs_(h5)
}
\arguments{
\item{h5}{The h5 file}
}
\description{
h5[['graphs']] file transform graphs information
}
