% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{to_layers_}
\alias{to_layers_}
\title{h5[['layers']] file transform the layer information}
\usage{
to_layers_(h5)
}
\arguments{
\item{h5}{The h5 file}
}
\description{
h5[['layers']] file transform the layer information
}
