% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIO.R
\docType{class}
\name{VisiumV1-class}
\alias{VisiumV1-class}
\alias{VisiumV1}
\title{create the VisiumV1 class inherting the package Seurat 'SpatialImage' class}
\description{
create the VisiumV1 class inherting the package Seurat 'SpatialImage' class
}
