% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIO.R
\name{read_h5part}
\alias{read_h5part}
\title{The read the some data in the h5 file}
\usage{
read_h5part(file, groups)
}
\arguments{
\item{file}{the h5 file}

\item{groups}{the h5 groups}
}
\value{
The list contains the core data for single-cell information
}
\description{
The read the some data in the h5 file
}
